+++ Assets/AccelByte/Models/Config.cs
@@ -12,6 +12,8 @@ namespace AccelByte.Models
         [DataMember] public string Namespace { get; set; }
         [DataMember] public bool UseSessionManagement { get; set; }
         [DataMember] public bool UsePlayerPrefs { get; set; }
+        [DataMember] public bool EnableDebugLog { get; set; }
+        [DataMember] public string DebugLogFilter { get; set; }
         [DataMember] public string BaseUrl { get; set; }
         [DataMember] public string ApiBaseUrl { get; set; }
         [DataMember] public string NonApiBaseUrl { get; set; }
@@ -49,6 +51,8 @@ namespace AccelByte.Models
             if (this.Namespace == anotherConfig.Namespace &&
                 this.UseSessionManagement == anotherConfig.UseSessionManagement &&
                 this.UsePlayerPrefs == anotherConfig.UsePlayerPrefs &&
+                this.EnableDebugLog == anotherConfig.EnableDebugLog &&
+                this.DebugLogFilter == anotherConfig.DebugLogFilter &&
                 this.BaseUrl == anotherConfig.BaseUrl &&
                 this.ApiBaseUrl == anotherConfig.ApiBaseUrl &&
                 this.NonApiBaseUrl == anotherConfig.NonApiBaseUrl &&